requirements.txt
pillow
django
crispy_bootstrap5
django-crispy-forms
flake8
black
python-dotenv
six

django-redis
djangorestframework-simplejwt
drf-spectacular
django-allauth[socialaccount]
django-jazzmin
pytest
pytest-django 
factory_boy 
Faker
pytest-cov
psycopg2-binary
google-auth
google-auth-oauthlib
google-api-python-client
pytest-xdist